Separate the two cases where mark_tree() was used by copying the implementation and unfolding the constants in the code directly.
The reason for this change is that in WC-NG both cases require
completely different handling and separating these cases allows
changing them one at a time. (The presence model changes how we
handle deleted and how we handle copied, so this is no longer
just switching a few flags as the old code did).

* subversion/libsvn_wc/adm_ops.c
  (mark_tree): Remove function.
  (mark_tree_deleted): New function, based on mark_tree()
  (svn_wc_delete4): Update caller.
  (mark_tree_copied): New function, based on mark_tree()
  (svn_wc_add4): Update caller.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@879701 13f79535-47bb-0310-9956-ffa450edef68