Fix issue #3519 - ra_serf using Label header unsupported by mod_dav_svn.
NOTE: This bug is triggered by switch_tests.py 19 on the 1.6.x branch.
It isn't exhibited in stock trunk testing because the HTTPv2 logic
doesn't attempt the kind of CHECKOUT request that tickles the problem.
And while it *should* show up in trunk with "SVNAdvertiseV2Protocol
off" set in httpd.conf, it currently doesn't because of an entirely
different bug (related to wcprops, and also currently causing ra_neon
to fail in the same way).  Confused?

* subversion/libsvn_ra_serf/commit.c
  (get_version_url): Don't call svn_ra_serf__deliver_props() in a way
    that would cause it to use a Label header (because mod_dav_svn
    doesn't do recognize the header at all).


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@880370 13f79535-47bb-0310-9956-ffa450edef68