Following up on r880496, in all tests that create their own config file, include the config option "password-stores=" which is present in the default config. Some tests already did so; these two did not.
The failure mode this prevents is if the dynamic auth libraries (gnome-keyring
and kwallet) fail to load, and if svn is compiled with SVN_DEBUG, then a
warning message is printed by svn_dso_load() on stderr which is seen by the
test suite as unexpected output and therefore a failure. The message is only
printed when compiled for debugging.

We may also want to remove the debug message, as an additional change.

* subversion/tests/cmdline/stat_tests.py
  (timestamp_behaviour): Add "password-stores=" to the config.

* subversion/tests/cmdline/update_tests.py
  (eof_in_interactive_conflict_resolver): Add "password-stores=" to the config.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@884648 13f79535-47bb-0310-9956-ffa450edef68