For obliterate in BDB, partially implement updating the "copies" table.
This creates the required new entries in the "copies" table when a new
obliteration txn is begun, and links them to the new txn, but this is not
sufficient because the new keys into the "copies" table then do not match
the copy-id component of the node-ids involved. This discrepancy needs to be
investigated and resolved.

* subversion/libsvn_fs_base/revs-txns.c
  (copy_dup): New function.
  (txn_body_begin_obliteration_txn): Call copy_dup(). Add a comment about
    issues to be resolved.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@890770 13f79535-47bb-0310-9956-ffa450edef68