Unify the writing of new WC-1 text-base files in three different places in update_editor.c.
This is a re-application of the patch I first committed in r919413 (and
Philip tweaked in r919436) and which I then reverted in r919478.  I have
since charted the usage of temp text bases (see
'notes/wc-ng/use-of-tmp-text-base-path') and found that the usage here is
independent of the special path given by svn_wc__text_base_path(), and is
safe to change.

* subversion/libsvn_wc/adm_files.h,
  subversion/libsvn_wc/adm_files.c
  (svn_wc__open_writable_base): Change from creating the temporary file at
    a special path (that depended on whether it was a revert-base) to
    creating it with an arbitrary and unique name in the admin temp dir.
  (open_adm_file): Remove.

* subversion/libsvn_wc/update_editor.c
  (add_file_with_history, svn_wc_add_repos_file4): Use
    svn_wc__open_writable_base() instead of the equivalent long-hand code.
  (apply_textdelta): Adjust by passing the 'db' instead of a 'revert' flag
    to svn_wc__open_writable_base().


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@926637 13f79535-47bb-0310-9956-ffa450edef68