Create a Windows specific implementation of file_mktemp() to work around access denied errors returned from apr's internal gettemp implementation. This moves the windows code back to an improved version of our 1.6 tempfile creation code.
The collision avoidance code I added in file_mktemp() is not really high tech,
but it doesn't cause a single collision when running the ra_local tests.
(Not sure if that is a good or a bad thing ;-)

The windows standard tempfile creation function doesn't support long path
names, so it can't be used in places where our temp functions are used.

* subversion/libsvn_subr/io.c
  (tempname_counter): New variable.
  (file_mktemp): Update argument to receive a directory path instead of a
    template and create the template for apr. Add a Windows specific
    implementation using some code borrowed from svn_io_open_uniquely_named().
  (svn_io_open_unique_file3): Update caller.

git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@944540 13f79535-47bb-0310-9956-ffa450edef68