Start matching property hunks.
With this change, we can handle matching of modified and deleted 
properties but we cannot yet handle added props. (The current code
for handling added files is not generic enough to apply for props).

Note that we do matching but none of the other steps necceseary for
installing a property patch.

* subversion/libsvn_client/patch.c
  (get_hunk_info): Add new parameter 'is_prop_hunk' for determining
    if match_new_file() should be called. The idea is that we should
    be able to handle matching of new props as well in the future.
    Make the match step more generic by checking if we have a target
    stream instead of as previously if our target was a file.
  (apply_one_patch): Match property hunks for each property in the
    same manner as is done for text hunks.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@966226 13f79535-47bb-0310-9956-ffa450edef68