Enable the apply step for properties.
With this change, we're writing the patched result to tmp files and
rejected hunks to another tmp file. The content of those files are
not yet installed, nor are we doing any notifications for the property
patching yet.

This patch has been tested by setting a breakpoint when the 'patched' 
streams for the properties are closed and reading the content of the
tmp files backing up those streams.

* subversion/libsvn_client/patch.c
  (reject_hunk,
   apply_hunk): Add 'is_prop_hunk' parameter to be able to separate
    the recording of rejects from those occuring for texts. For now, 
    we've disabled recording rejects for properties. Let's do things
    one step at a time.
  (apply_one_patch): For each property, write it's hunks to the patched stream
    or the reject stream by calling apply_hunk() or reject_hunk() in the
    same manner as is done for file content.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@966403 13f79535-47bb-0310-9956-ffa450edef68