Follow-up to r982534. Remove obsolute TODO's.
We're not going to do any checks to compare the paths found 
on the '--git diff a/old_path b/new_path' line with paths found 
in subsequent lines.  I haven't come up with a way to handle 
paths with spaces in the git diff line unless old_path and 
new_path are the same which is only guarenteed for added and
deleted empty paths.

* subversion/libsvn_diff/parse-diff.c
  (git_minus,
   git_plus,
   git_move_from,
   git_move_to,
   git_copy_from,
   git_copy_to,
   svn_diff_parse_next_patch): Remove obsolute TODO's.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@982563 13f79535-47bb-0310-9956-ffa450edef68