Resolve a regression in the working copy status walker that made missing directories invisible when using 'svn status -u' in single-db mode.
Before single-db we reported missing directories as deleted to the
repository in the crawler step to receive them as changes. But we
stopped doing that as we still have all their information. But when
walking the status we still skipped these paths from their parent
directory as we assumed that they would be reported via the repository
status.

This patch allows fixing merge_tests.py 16 in a followup commit.

* subversion/libsvn_wc/status.c
  (make_dir_baton,
   handle_statii): Don't skip missing or obstructed directories when
     checking children in single-db.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@990798 13f79535-47bb-0310-9956-ffa450edef68