Fix revert_tests.py 21 by making svn_wc_revert4 check if a revert is possible with only impact at the specified depth.
Before WC-NG libsvn_client didn't take a recursive lock if the specified
depth was less than infinity and some revert operations then failed on
not having a lock. But this didn't fix that you could do a depth empty
revert on a directory which then removes an added (child-of-copy) file.

As far as I can see there is no sane way to keep the original behavior or
reimplement the old behavior inside one of the deprecated wrappers.

svn_wc_revert3 receives an access baton and a depth and failed the revert
if the lock in the access baton wasn't deep enough. (Not on a depth value).

svn_client_revert2 doesn't document it's error behavior, so I think it is
safe to return a different error here, but this might need a later update
to wrap the error with SVN_ERR_WC_NOT_LOCKED to fix clients that handle
this error like we did in svn before this patch.

See also
http://svn.haxx.se/dev/archive-2010-08/0763.shtml

* subversion/include/svn_error_codes.h
  (SVN_ERR_WC_INVALID_OPERATION_DEPTH): New error code.

* subversion/libsvn_wc/adm_ops.c
  (verify_revert_depth): New function.
  (revert_internal): Add depth verification step.

* subversion/svn/revert-cmd.c
  (svn_cl__revert): Update error code check.

* subversion/tests/cmdline/revert_tests.py
  (test_list): Remove XFail marking from revert_added_tree.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@991158 13f79535-47bb-0310-9956-ffa450edef68