Enable empty files to be added with svn patch.
This only works if the diff uses the git extension format. The extra
logic is for allowing us to create added empty files for paths that
has property changes, be the patch in the git format or not. If we only allowed
git diffs to create empty files, the code could be simplified.

* subversion/libsvn_client/patch.c
  (init_patch_target): Mark the target as added if the patch has recorded 
    it as <to be added>.
  (apply_one_patch): Don't skip paths that should be added according to the
    git diff information.
  (install_patched_prop_targets): Check if the path is already added before
    trying to add that was not existing when the patching process began.
  (apply_patches): Allow added paths to be installed even if they are empty.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@991508 13f79535-47bb-0310-9956-ffa450edef68