Merge r985037, r985046, r995507 and r995603 from the performance branch.
These changes introduce the svn_stringbuf_appendbyte() function, which has
significantly less overhead than svn_stringbuf_appendbytes(), and can be
used in a number of places within our codebase.

r995507 is the removal of an extra function on the branch.  This function
conflicted with previous revisions in the merge, and is itself no longer on
trunk.


git-svn-id: http://svn.apache.org/repos/asf/subversion/trunk@997203 13f79535-47bb-0310-9956-ffa450edef68